Added routers into the project.  also added a new function allowing us to create new users.
